+++ sshd-core/src/test/java/org/apache/sshd/util/test/EchoCommand.java
@@ -16,6 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+
 package org.apache.sshd.util.test;
 
 import java.io.OutputStream;
@@ -24,17 +25,16 @@ import java.nio.charset.StandardCharsets;
 /**
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-public class EchoShell extends CommandExecutionHelper {
-    public EchoShell() {
-        super();
+public class EchoCommand extends CommandExecutionHelper {
+    public EchoCommand(String command) {
+        super(command);
     }
 
     @Override
     protected boolean handleCommandLine(String command) throws Exception {
         OutputStream out = getOutputStream();
-        out.write((command + "\n").getBytes(StandardCharsets.UTF_8));
+        out.write(command.getBytes(StandardCharsets.UTF_8));
         out.flush();
-
-        return !"exit".equals(command);
+        return true;
     }
 }